Testes para fetchActivityDetails
Adiciona testes para fetchActivityDetails;
Corrige conflitos;
Movidos códigos de arquivos repetidos em um só;
