+++ solutions/devsprint-cecilia-saraiva-2/FinanceApp/Service/Entities/ActivityDetails.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct ActivityDetails: Decodable {
+struct ActivityDetails: Codable, Equatable {
 
     let name: String
     let price: Float
