+++ solutions/devsprint-cecilia-saraiva-2/FinanceAppTests/Mock/JSON/ActivityDetailsInvalid.json
@@ -1,0 +1,6 @@
+{
+    "name": "Marvel",
+    "price": "10.00",
+    "category": "Movie",
+    "time": "21:00"
+}
